requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==4.0.0
    # via jupyter-server
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.2.3
    # via isoduration
asttokens==2.4.0
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.12.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.2
    # via nbconvert
bleach==6.0.0
    # via nbconvert
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via pip-tools
cmdstanpy==1.1.0
    # via gptools-stan
colorama==0.4.6
    # via cook-build
comm==0.1.4
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.1.0
    # via matplotlib
cook-build==0.5.1
    # via -r requirements.in
cycler==0.11.0
    # via matplotlib
debugpy==1.8.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
exceptiongroup==1.1.3
    # via
    #   anyio
    #   ipython
executing==1.2.0
    # via stack-data
fastjsonschema==2.18.0
    # via nbformat
fonttools==4.42.1
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
gptools-stan==0.1.4
    # via -r requirements.in
gptools-util==0.1.1
    # via
    #   -r requirements.in
    #   gptools-stan
idna==3.4
    # via
    #   anyio
    #   jsonschema
    #   requests
ipykernel==6.25.2
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
    #   qtconsole
ipython==8.15.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via qtconsole
ipywidgets==8.1.0
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.0
    # via ipython
jinja2==3.1.2
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.9.14
    # via jupyterlab-server
jsonpointer==2.4
    # via jsonschema
jsonschema[format-nongpl]==4.19.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.7.1
    # via jsonschema
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==8.3.1
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
    #   qtconsole
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.3.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
    #   qtconsole
jupyter-events==0.7.0
    # via jupyter-server
jupyter-lsp==2.2.0
    # via jupyterlab
jupyter-server==2.7.3
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.4.4
    # via jupyter-server
jupyterlab==4.0.5
    # via notebook
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-server==2.25.0
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.8
    # via ipywidgets
jupytext==1.15.1
    # via -r requirements.in
kiwisolver==1.4.5
    # via matplotlib
markdown-it-py==3.0.0
    # via
    #   jupytext
    #   mdit-py-plugins
markupsafe==2.1.3
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.7.3
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mdit-py-plugins==0.4.0
    # via jupytext
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.1
    # via nbconvert
nbclient==0.8.0
    # via nbconvert
nbconvert==7.8.0
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.9.2
    # via
    #   jupyter-server
    #   jupytext
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.7
    # via ipykernel
networkx==3.1
    # via cook-build
notebook==7.0.3
    # via jupyter
notebook-shim==0.2.3
    # via
    #   jupyterlab
    #   notebook
numpy==1.25.2
    # via
    #   cmdstanpy
    #   contourpy
    #   gptools-stan
    #   gptools-util
    #   matplotlib
    #   pandas
    #   scipy
overrides==7.4.0
    # via jupyter-server
packaging==23.1
    # via
    #   build
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
    #   qtconsole
    #   qtpy
pandas==2.1.0
    # via cmdstanpy
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==10.0.0
    # via matplotlib
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.10.0
    # via jupyter-core
prometheus-client==0.17.1
    # via jupyter-server
prompt-toolkit==3.0.39
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.16.1
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyparsing==3.1.1
    # via matplotlib
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
    #   pandas
python-json-logger==2.0.7
    # via jupyter-events
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   jupyter-events
    #   jupytext
pyzmq==25.1.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   qtconsole
qtconsole==5.4.4
    # via jupyter
qtpy==2.4.0
    # via qtconsole
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.31.0
    # via jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.10.2
    # via
    #   jsonschema
    #   referencing
scipy==1.11.2
    # via -r requirements.in
send2trash==1.8.2
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.0
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
stack-data==0.6.2
    # via ipython
tabulate==0.9.0
    # via -r requirements.in
terminado==0.17.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.2.1
    # via nbconvert
toml==0.10.2
    # via jupytext
tomli==2.0.1
    # via
    #   build
    #   jupyterlab
    #   pip-tools
    #   pyproject-hooks
tornado==6.3.3
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
tqdm==4.66.1
    # via cmdstanpy
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   qtconsole
typing-extensions==4.7.1
    # via async-lru
tzdata==2023.3
    # via pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.0.4
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.6.3
    # via jupyter-server
wheel==0.41.2
    # via pip-tools
widgetsnbextension==4.0.8
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
